Exceptions de negócio

+++ src/main/java/com/example/devsuperior/client/service/ClientService.java
@@ -2,9 +2,11 @@ package com.example.devsuperior.client.service;
 
 import com.example.devsuperior.client.dto.ClientResponseDTO;
 import com.example.devsuperior.client.dto.ClientRequestDTO;
+import com.example.devsuperior.client.exception.ResourceNotFoundException;
 import com.example.devsuperior.client.model.Client;
 import com.example.devsuperior.client.repository.ClientRepository;
 import org.modelmapper.ModelMapper;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -23,7 +25,8 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientResponseDTO findById(Long id) {
-        Client client = repository.findById(id).get();
+        Client client = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Cliente não encontrado" +
+                " com o id: " + id));
 
         return modelMapper.map(client, ClientResponseDTO.class);
     }
@@ -37,6 +40,10 @@ public class ClientService {
 
     @Transactional
     public ClientResponseDTO insert(ClientRequestDTO dto) {
+        if(repository.existsByCpf(dto.getCpf())) {
+            throw new DataIntegrityViolationException("O cpf " + dto.getCpf() + " já está sendo usado por outro usuário.");
+        }
+
         Client client = modelMapper.map(dto, Client.class);
 
         client = repository.save(client);
@@ -46,7 +53,15 @@ public class ClientService {
 
     @Transactional
     public ClientResponseDTO update(Long id, ClientRequestDTO dto) {
-        Client client = repository.findById(id).get();
+        Client client = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Cliente não encontrado" +
+                " com o id: " + id));
+
+        //Valida se o cpf é diferente do que está salvo. Se for,verifica se outro cliente tem o cpf cadastrado
+        if (!dto.getCpf().equals(client.getCpf())) {
+            if (repository.existsByCpf(dto.getCpf())) {
+                throw new DataIntegrityViolationException("Não é possível atualizar. O cpf " + dto.getCpf() + " já está em uso.");
+            }
+        }
 
         modelMapper.map(dto, client);
 
@@ -57,6 +72,10 @@ public class ClientService {
 
     @Transactional
     public void delete(Long id){
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Não foi possível deletar. Cliente não encontrado " +
+                    "com o ID:" + id);
+        }
         repository.deleteById(id);
     }
 }
